Refactor - wip
