+++ cluster/FailsafeTask.java
@@ -19,7 +19,7 @@
  * under the License.
  */
 
-package com.vaticle.typedb.client.rpc.cluster;
+package com.vaticle.typedb.client.cluster;
 
 import com.vaticle.typedb.client.common.exception.TypeDBClientException;
 import com.vaticle.typedb.protocol.ClusterDatabaseProto;
