+++ core/CoreClient.java
@@ -19,7 +19,7 @@
  * under the License.
  */
 
-package com.vaticle.typedb.client.rpc.core;
+package com.vaticle.typedb.client.core;
 
 import com.vaticle.typedb.client.rpc.TypeDBClientAbstract;
 import com.vaticle.typedb.client.rpc.TypeDBConnectionFactory;
